Flatten a branches by negation
So instead of doing `if (a) {...}`, we do `if (!a) continue; ...`

This way we have less nesting and more clear branching for if statements; if a loop/generator/function is going to break/continue/return, this will be done early (i.e. in the if statement rather than else or otherwise).