Add lots of languages to genre/language check
Doesn't matter that the website doesn't have them, as they are mutually exclusive anyway. If any of these exist, the web language can only be "Other" in the vast majority of cases.

Only exception I can think of is if there are multiple languages in the song and none of the web ones are in the tags but others are.